Remove duplicate defines that cause non-deterministic behavior
Some type constants copied from cvinfo.h are not unique. In C this is
not a problem since code will only refer to the named constant, but in
Python they end up stored in a dict that is indexed by value rather than
name and as a result the contents of the dict depend on the order in
which the keys get added, which is nondeterministic in Python 3.3.

Fix kindly provided by @lucasg.